+++ code/master/api/tethys_api/urls.py
@@ -35,6 +35,9 @@ urlpatterns = [
     path('api/channel/<str:number>', views.channel_single),
     path('api/channel/<str:number>/<str:action>', views.channel_single_action),
 
+    path('api/manualCommand/pending', views.manualCommand_pending),
+    path('api/manualCommand/<int:id>', views.manualCommand_single),
+
     path('api/channelSummary/', views.channelSummary),
     path('api/channelSummary/<int:number>', views.channelSummary_single),